Exclude draft posts from sitemap
Resolves #135
