+++ package.json
@@ -48,6 +48,7 @@
         "@typescript-eslint/typescript-estree": "^8.50.1",
         "prettier": "^3.7.4",
         "prettier-plugin-astro": "^0.14.1",
-        "prettier-plugin-tailwindcss": "^0.7.1"
+        "prettier-plugin-tailwindcss": "^0.7.1",
+        "vite": "^7.3.1"
     }
 }
