+++ src/lib/postsLoader.ts
@@ -1,10 +1,6 @@
 import type { Loader, LoaderContext } from "astro/loaders";
 import { z } from "astro:content";
-
-const convertToDate = (dateString: string): Date =>
-    new Date(
-        dateString + (!dateString.includes("T") ? "T07:00:00.000-08:00" : ""),
-    );
+import { getPosts } from "./posts";
 
 // TODO: Change `reorganize` to `main`
 const loader = (options: { apiKey: string }): Loader => {
@@ -39,63 +35,7 @@ const loader = (options: { apiKey: string }): Loader => {
             parseData,
             renderMarkdown,
         }: LoaderContext): Promise<void> => {
-            const response = await fetch(
-                "https://raw.githubusercontent.com/hkamran80/articles/reorganize/index.json",
-            );
-            // TODO: Add actual type
-            const data = (await response.json()) as any[];
-
-            const posts = data.map((post) => {
-                let converted: { [key: string]: string | Date | object } = {};
-
-                if (post.status !== "draft") {
-                    converted.published = convertToDate(post.published);
-
-                    if (post.type === "article")
-                        converted.img = {
-                            src: `https://assets.hkamran.com/graphics/article/${post.id}`,
-                            alt: post.imgAlt,
-                        };
-                }
-
-                if (post.status === "updated")
-                    converted.updated = convertToDate(post.updated);
-
-                return {
-                    ...post,
-                    ...converted,
-                };
-            });
-
-            const branchResponse = await fetch(
-                "https://api.github.com/repos/hkamran80/articles/branches",
-                {
-                    headers: {
-                        Accept: "application/vnd.github+json",
-                        "X-GitHub-Api-Version": "2022-11-28",
-                        Authorization: `Bearer ${options.apiKey}`,
-                    },
-                },
-            );
-            const branches = (await branchResponse.json()) as {
-                name: string;
-            }[];
-            for (const branch of branches) {
-                if (
-                    branch.name.startsWith("article-r/") ||
-                    branch.name.startsWith("note-r/") ||
-                    branch.name.startsWith("post/")
-                ) {
-                    const response = await fetch(
-                        `https://raw.githubusercontent.com/hkamran80/articles/${branch.name}/index.json`,
-                    );
-                    // TODO: Add actual type
-                    const data = (await response.json()) as any[];
-
-                    const postId = branch.name.split("/")[1];
-                    posts.push(data.find(({ id }) => id === postId));
-                }
-            }
+            const posts = await getPosts(options.apiKey);
 
             for (const post of posts) {
                 const data = await parseData({
